feat: implement LogStore with MemoryStore backend
Add centralized log storage system to enable HTTP API access for LLM tools
while maintaining TUI performance through direct store reads.

Changes:
- Add LogStore trait with core methods (append, get_lines, get_lines_since,
  get_process, get_processes)
- Implement MemoryStore with ring buffer per-process log storage
- Update ProcessInfo to generate stable process IDs via hash of (name, cmd, cwd)
- Add ProcessInfo::new() constructor with encapsulated ID computation
- Refactor StoreEvent to use process_id (u64) and ProcessStatus enum
- Update LogLine to include process_id, session_id, and timestamp (u128)
- Add SessionInfo struct for session tracking
- Refactor Supervisor to use ProcessContext struct wrapping ProcessInfo
- Update Process to emit StoreEvent::AppendLog with separate stdout/stderr
- Wire ProcessInfo through server::serve() and main.rs

Implementation details:
- Single HashMap<u64, ProcessData> with pre-allocated VecDeque ring buffers
- LogLine IDs derived from last entry, timestamps generated internally
- Binary search for get_lines_since() with graceful eviction handling
- Pre-allocation in MemoryStore::new() for known process list

+++ cdi-shared/src/log.rs
@@ -9,14 +9,14 @@ pub struct LogLine {
     pub id: u64,
     pub process_id: u64,
     pub session_id: u64,
-    pub timestamp: u64,
+    pub timestamp: u128,
     pub stream: Stream,
     pub content: String,
 }
 
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub enum ProcessStatus {
-    Running, 
+    Running,
     Stopped,
     Crashed,
 }
@@ -48,9 +48,9 @@ impl ProcessInfo {
     }
 
     fn compute_id(name: &str, command: &str, cwd: Option<&str>) -> u64 {
-        use std::hash::{Hash, Hasher};
         use std::collections::hash_map::DefaultHasher;
-        
+        use std::hash::{Hash, Hasher};
+
         let mut hasher = DefaultHasher::new();
         name.hash(&mut hasher);
         command.hash(&mut hasher);
@@ -58,7 +58,6 @@ impl ProcessInfo {
         hasher.finish()
     }
 }
-
 
 #[derive(Clone, Debug)]
 pub struct SessionInfo {
